requirements.txt
absl-py==0.1.13
astor==0.6.2
bleach==1.5.0
click==6.7
cycler==0.10.0
Cython==0.28.1
decorator==4.2.1
dlib==19.10.0
face-recognition-models==0.3.0
gast==0.2.0
gdbn==0.1
gnumpy==0.2
google-images-download==2.1.2
grpcio==1.10.0
h5py==2.7.1
html5lib==0.9999999
imutils==0.4.6
joblib==0.11
Keras==2.1.6
kiwisolver==1.0.1
Lasagne==0.1
Markdown==2.6.11
matplotlib==2.2.2
networkx==2.1
nolearn==0.6.0
numpy==1.14.2
opencv-python==3.4.0.12
pafy==0.5.4
pandas==0.22.0
Pillow==5.1.0
protobuf==3.5.2.post1
pygame==1.9.3
pyparsing==2.2.0
python-dateutil==2.7.2
pytz==2018.3
PyWavelets==0.5.2
PyYAML==3.12
scikit-image==0.13.1
scikit-learn==0.19.1
scipy==1.0.1
selenium==3.12.0
six==1.11.0
sklearn==0.0
tabulate==0.8.2
tensorboard==1.7.0
tensorflow==1.7.0
termcolor==1.1.0
Theano==1.0.1
Werkzeug==0.14.1
youtube-dl==2018.5.9
bcdedit /set hypervisorlaunchtype off